chore: :shirt:

+++ spec/certora/contracts/mcd/Pot.sol
@@ -39,32 +39,48 @@ import "./lib.sol";
 */
 
 contract VatLike {
-    function move(address,address,uint256) external;
-    function suck(address,address,uint256) external;
+    function move(
+        address,
+        address,
+        uint256
+    ) external;
+
+    function suck(
+        address,
+        address,
+        uint256
+    ) external;
 }
 
 contract Pot is LibNote {
     // --- Auth ---
-    mapping (address => uint) public wards;
-    function rely(address guy) external note auth { wards[guy] = 1; }
-    function deny(address guy) external note auth { wards[guy] = 0; }
-    modifier auth {
+    mapping(address => uint256) public wards;
+
+    function rely(address guy) external note auth {
+        wards[guy] = 1;
+    }
+
+    function deny(address guy) external note auth {
+        wards[guy] = 0;
+    }
+
+    modifier auth() {
         require(wards[msg.sender] == 1, "Pot/not-authorized");
         _;
     }
 
     // --- Data ---
-    mapping (address => uint256) public pie;  // user Savings Dai
+    mapping(address => uint256) public pie; // user Savings Dai
 
-    uint256 public Pie;  // total Savings Dai
-    uint256 public dsr;  // the Dai Savings Rate
-    uint256 public chi;  // the Rate Accumulator
+    uint256 public Pie; // total Savings Dai
+    uint256 public dsr; // the Dai Savings Rate
+    uint256 public chi; // the Rate Accumulator
 
-    VatLike public vat;  // CDP engine
-    address public vow;  // debt engine
-    uint256 public rho;  // time of last drip
+    VatLike public vat; // CDP engine
+    address public vow; // debt engine
+    uint256 public rho; // time of last drip
 
-    uint256 public live;  // Access Flag
+    uint256 public live; // Access Flag
 
     // --- Init ---
     constructor(address vat_) public {
@@ -77,24 +93,56 @@ contract Pot is LibNote {
     }
 
     // --- Math ---
-    uint256 constant ONE = 10 ** 27;
-    function rpow(uint x, uint n, uint base) internal pure returns (uint z) {
+    uint256 constant ONE = 10**27;
+
+    function rpow(
+        uint256 x,
+        uint256 n,
+        uint256 base
+    ) internal pure returns (uint256 z) {
         assembly {
-            switch x case 0 {switch n case 0 {z := base} default {z := 0}}
+            switch x
+            case 0 {
+                switch n
+                case 0 {
+                    z := base
+                }
+                default {
+                    z := 0
+                }
+            }
             default {
-                switch mod(n, 2) case 0 { z := base } default { z := x }
-                let half := div(base, 2)  // for rounding.
-                for { n := div(n, 2) } n { n := div(n,2) } {
+                switch mod(n, 2)
+                case 0 {
+                    z := base
+                }
+                default {
+                    z := x
+                }
+                let half := div(base, 2) // for rounding.
+                for {
+                    n := div(n, 2)
+                } n {
+                    n := div(n, 2)
+                } {
                     let xx := mul(x, x)
-                    if iszero(eq(div(xx, x), x)) { revert(0,0) }
+                    if iszero(eq(div(xx, x), x)) {
+                        revert(0, 0)
+                    }
                     let xxRound := add(xx, half)
-                    if lt(xxRound, xx) { revert(0,0) }
+                    if lt(xxRound, xx) {
+                        revert(0, 0)
+                    }
                     x := div(xxRound, base)
-                    if mod(n,2) {
+                    if mod(n, 2) {
                         let zx := mul(z, x)
-                        if and(iszero(iszero(x)), iszero(eq(div(zx, x), z))) { revert(0,0) }
+                        if and(iszero(iszero(x)), iszero(eq(div(zx, x), z))) {
+                            revert(0, 0)
+                        }
                         let zxRound := add(zx, half)
-                        if lt(zxRound, zx) { revert(0,0) }
+                        if lt(zxRound, zx) {
+                            revert(0, 0)
+                        }
                         z := div(zxRound, base)
                     }
                 }
@@ -102,19 +150,19 @@ contract Pot is LibNote {
         }
     }
 
-    function rmul(uint x, uint y) internal pure returns (uint z) {
+    function rmul(uint256 x, uint256 y) internal pure returns (uint256 z) {
         z = mul(x, y) / ONE;
     }
 
-    function add(uint x, uint y) internal pure returns (uint z) {
+    function add(uint256 x, uint256 y) internal pure returns (uint256 z) {
         require((z = x + y) >= x);
     }
 
-    function sub(uint x, uint y) internal pure returns (uint z) {
+    function sub(uint256 x, uint256 y) internal pure returns (uint256 z) {
         require((z = x - y) <= x);
     }
 
-    function mul(uint x, uint y) internal pure returns (uint z) {
+    function mul(uint256 x, uint256 y) internal pure returns (uint256 z) {
         require(y == 0 || (z = x * y) / y == x);
     }
 
@@ -137,26 +185,26 @@ contract Pot is LibNote {
     }
 
     // --- Savings Rate Accumulation ---
-    function drip() external note returns (uint tmp) {
+    function drip() external note returns (uint256 tmp) {
         require(now >= rho, "Pot/invalid-now");
         tmp = rmul(rpow(dsr, now - rho, ONE), chi);
-        uint chi_ = sub(tmp, chi);
+        uint256 chi_ = sub(tmp, chi);
         chi = tmp;
         rho = now;
         vat.suck(address(vow), address(this), mul(Pie, chi_));
     }
 
     // --- Savings Dai Management ---
-    function join(uint wad) external note {
+    function join(uint256 wad) external note {
         require(now == rho, "Pot/rho-not-updated");
         pie[msg.sender] = add(pie[msg.sender], wad);
-        Pie             = add(Pie,             wad);
+        Pie = add(Pie, wad);
         vat.move(msg.sender, address(this), mul(chi, wad));
     }
 
-    function exit(uint wad) external note {
+    function exit(uint256 wad) external note {
         pie[msg.sender] = sub(pie[msg.sender], wad);
-        Pie             = sub(Pie,             wad);
+        Pie = sub(Pie, wad);
         vat.move(address(this), msg.sender, mul(chi, wad));
     }
 }
